#822: dedup typst backend + PDF crate tests (~280 lines)
Remove redundant/low-value tests flagged by the minimize review, incl. the
items it marked as judgment calls.

- Delete tests/eval_error_hint.rs (retired eval-hint path); move its one live
  test, resolvable_eval_error_is_unchanged, into error_mapping.rs's fixture
  test module (sharing a walk_fixture/source_with_plate helper).
- Delete the test_severity_mapping tautology (never calls map_single_diagnostic).
- sig_field.rs: delete regression_widget_dict_has_exactly_one_subtype (the
  double-/Subtype byte-check is owned by quillmark-pdf/tests/stamp.rs); drop the
  spine-emission re-asserts (MULTILINE/COMBO Ff bits, /MK /CA glyph) and the
  orphaned field_flags helper, keeping the adapter-mapping asserts (/V binding,
  checkbox truthiness, choice option matching). Fold in the "bound widget /T
  name must not leak as a region" assertion.
- content_regions.rs: merge the byte-identical field_placed_twice / field_at_
  resolves_every_placement pair into one session; delete
  form_field_region_needs_a_schema_binding (dup of the sig_field test).
- quillmark-pdf/tests/stamp.rs: build_base_pdf delegates to build_base_pdf_origin.
- pdfform/tests/sample_form.rs: trim spine bit-flag re-asserts already pinned at
  the stamp seam, keeping the binding-layer coverage (values, tooltip, array
  join, regions, producer default).
- pdfform/src/flatten.rs: fold flatten_has_fonts_and_text_operators down to the
  one /WinAnsiEncoding check the byte-window + lopdf tests don't already cover.
